requirements.txt
# Testing
pytest>=8.0.0

# API Integration
requests>=2.31.0
aiohttp>=3.9.0
vt-py>=0.18.0
python-dotenv>=1.0.0

# Heuristic Analysis
pefile>=2023.2.7
python-magic-bin>=0.4.14
numpy>=1.26.0
scikit-learn>=1.3.0
yara-python>=4.5.0

# Behavior Monitoring
watchdog>=3.0.0
psutil>=5.9.0
pywin32>=306

# GUI
customtkinter>=5.2.0
pillow>=10.1.0
matplotlib>=3.8.0

# Utilities
pyyaml>=6.0.1
apscheduler>=3.10.0
cryptography>=41.0.0
reportlab>=4.0.0
